moved class flag to flagParam
